+++ app/src/main/java/com/ganaljigi/kubf/ui/roominfo/di/RoomInfoApiModule.kt
@@ -1,5 +1,6 @@
-package com.ganaljigi.kubf.ui.roominfo
+package com.ganaljigi.kubf.ui.roominfo.di
 
+import com.ganaljigi.kubf.ui.roominfo.service.RoomInfoService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
