+++ app/src/main/java/com/ganaljigi/kubf/ui/roominfo/di/RoomInfoRepositoryModule.kt
@@ -1,5 +1,7 @@
-package com.ganaljigi.kubf.ui.roominfo
+package com.ganaljigi.kubf.ui.roominfo.di
 
+import com.ganaljigi.kubf.ui.roominfo.repository.RoomInfoRepository
+import com.ganaljigi.kubf.ui.roominfo.repositoryimpl.RoomInfoRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
