+++ app/src/main/java/com/ganaljigi/kubf/ui/roominfo/repository/RoomInfoRepository.kt
@@ -1,4 +1,6 @@
-package com.ganaljigi.kubf.ui.roominfo
+package com.ganaljigi.kubf.ui.roominfo.repository
+
+import com.ganaljigi.kubf.ui.roominfo.response.RoomInfoResponseDto
 
 interface RoomInfoRepository {
     suspend fun getRoomInfo(
